requirements.txt
pandas
numpy
scikit-learn
joblib
fastapi
uvicorn
firebase-admin
python-dotenv
pydantic
python-multipart
